Workfile Templates: Implement registering and discovering of `PlaceholderPlugin`
