+++ client/ayon_core/hosts/aftereffects/api/workfile_template_builder.py
@@ -1,6 +1,7 @@
 import os.path
 import uuid
 import shutil
+from abc import abstractmethod
 
 from ayon_core.pipeline import registered_host
 from ayon_core.tools.workfile_template_build import (
@@ -9,13 +10,9 @@ from ayon_core.tools.workfile_template_build import (
 from ayon_core.pipeline.workfile.workfile_template_builder import (
     AbstractTemplateBuilder,
     PlaceholderPlugin,
-    LoadPlaceholderItem,
-    CreatePlaceholderItem,
-    PlaceholderLoadMixin,
-    PlaceholderCreateMixin
+    PlaceholderItem
 )
 from ayon_core.hosts.aftereffects.api import get_stub
-from ayon_core.hosts.aftereffects.api.lib import set_settings
 
 PLACEHOLDER_SET = "PLACEHOLDERS_SET"
 PLACEHOLDER_ID = "openpype.placeholder"
@@ -51,6 +48,10 @@ class AETemplateBuilder(AbstractTemplateBuilder):
 class AEPlaceholderPlugin(PlaceholderPlugin):
     """Contains generic methods for all PlaceholderPlugins."""
 
+    @abstractmethod
+    def _create_placeholder_item(self, item_data: dict) -> PlaceholderItem:
+        pass
+
     def collect_placeholders(self):
         """Collect info from file metadata about created placeholders.
 
@@ -63,17 +64,7 @@ class AEPlaceholderPlugin(PlaceholderPlugin):
             if item.get("plugin_identifier") != self.identifier:
                 continue
 
-            if isinstance(self, AEPlaceholderLoadPlugin):
-                item = LoadPlaceholderItem(item["uuid"],
-                                           item["data"],
-                                           self)
-            elif isinstance(self, AEPlaceholderCreatePlugin):
-                item = CreatePlaceholderItem(item["uuid"],
-                                             item["data"],
-                                             self)
-            else:
-                raise NotImplementedError(f"Not implemented for {type(self)}")
-
+            item = self._create_placeholder_item(item)
             output.append(item)
 
         return output
@@ -133,87 +124,6 @@ class AEPlaceholderPlugin(PlaceholderPlugin):
             "members": [item_id]
         }
         stub.imprint(item_id, container_data)
-
-
-class AEPlaceholderCreatePlugin(AEPlaceholderPlugin, PlaceholderCreateMixin):
-    """Adds Create placeholder.
-
-    This adds composition and runs Create
-    """
-    identifier = "aftereffects.create"
-    label = "AfterEffects create"
-
-    def create_placeholder(self, placeholder_data):
-        stub = get_stub()
-        name = "CREATEPLACEHOLDER"
-        item_id = stub.add_item(name, "COMP")
-
-        self._imprint_item(item_id, name, placeholder_data, stub)
-
-    def populate_placeholder(self, placeholder):
-        """Replace 'placeholder' with publishable instance.
-
-        Renames prepared composition name, creates publishable instance, sets
-        frame/duration settings according to DB.
-        """
-        pre_create_data = {"use_selection": True}
-        item_id, item = self._get_item(placeholder)
-        get_stub().select_items([item_id])
-        self.populate_create_placeholder(placeholder, pre_create_data)
-
-        # apply settings for populated composition
-        item_id, metadata_item = self._get_item(placeholder)
-        set_settings(True, True, [item_id])
-
-    def get_placeholder_options(self, options=None):
-        return self.get_create_plugin_options(options)
-
-
-class AEPlaceholderLoadPlugin(AEPlaceholderPlugin, PlaceholderLoadMixin):
-    identifier = "aftereffects.load"
-    label = "AfterEffects load"
-
-    def create_placeholder(self, placeholder_data):
-        """Creates AE's Placeholder item in Project items list.
-
-         Sets dummy resolution/duration/fps settings, will be replaced when
-         populated.
-         """
-        stub = get_stub()
-        name = "LOADERPLACEHOLDER"
-        item_id = stub.add_placeholder(name, 1920, 1060, 25, 10)
-
-        self._imprint_item(item_id, name, placeholder_data, stub)
-
-    def populate_placeholder(self, placeholder):
-        """Use Openpype Loader from `placeholder` to create new FootageItems
-
-        New FootageItems are created, files are imported.
-        """
-        self.populate_load_placeholder(placeholder)
-        errors = placeholder.get_errors()
-        stub = get_stub()
-        if errors:
-            stub.print_msg("\n".join(errors))
-        else:
-            if not placeholder.data["keep_placeholder"]:
-                metadata = stub.get_metadata()
-                for item in metadata:
-                    if not item.get("is_placeholder"):
-                        continue
-                    scene_identifier = item.get("uuid")
-                    if (scene_identifier and
-                            scene_identifier == placeholder.scene_identifier):
-                        stub.delete_item(item["members"][0])
-                stub.remove_instance(placeholder.scene_identifier, metadata)
-
-    def get_placeholder_options(self, options=None):
-        return self.get_load_plugin_options(options)
-
-    def load_succeed(self, placeholder, container):
-        placeholder_item_id, _ = self._get_item(placeholder)
-        item_id = container.id
-        get_stub().add_item_instead_placeholder(placeholder_item_id, item_id)
 
 
 def build_workfile_template(*args, **kwargs):
