+++ client/ayon_core/hosts/nuke/api/pipeline.py
@@ -18,6 +18,7 @@ from ayon_core.pipeline import (
     register_loader_plugin_path,
     register_creator_plugin_path,
     register_inventory_action_path,
+    register_template_placeholder_plugin_path,
     AYON_INSTANCE_ID,
     AVALON_INSTANCE_ID,
     AVALON_CONTAINER_ID,
@@ -52,8 +53,6 @@ from .lib import (
     MENU_LABEL,
 )
 from .workfile_template_builder import (
-    NukePlaceholderLoadPlugin,
-    NukePlaceholderCreatePlugin,
     build_workfile_template,
     create_placeholder,
     update_placeholder,
@@ -76,6 +75,7 @@ PUBLISH_PATH = os.path.join(PLUGINS_DIR, "publish")
 LOAD_PATH = os.path.join(PLUGINS_DIR, "load")
 CREATE_PATH = os.path.join(PLUGINS_DIR, "create")
 INVENTORY_PATH = os.path.join(PLUGINS_DIR, "inventory")
+TEMPLATE_PLUGINS_PATH = os.path.join(PLUGINS_DIR, "template")
 
 # registering pyblish gui regarding settings in presets
 if os.getenv("PYBLISH_GUI", None):
@@ -105,18 +105,11 @@ class NukeHost(
     def get_workfile_extensions(self):
         return file_extensions()
 
-    def get_workfile_build_placeholder_plugins(self):
-        return [
-            NukePlaceholderLoadPlugin,
-            NukePlaceholderCreatePlugin
-        ]
-
     def get_containers(self):
         return ls()
 
     def install(self):
-        ''' Installing all requarements for Nuke host
-        '''
+        """Installing all requirements for Nuke host"""
 
         pyblish.api.register_host("nuke")
 
@@ -125,6 +118,7 @@ class NukeHost(
         register_loader_plugin_path(LOAD_PATH)
         register_creator_plugin_path(CREATE_PATH)
         register_inventory_action_path(INVENTORY_PATH)
+        register_template_placeholder_plugin_path(TEMPLATE_PLUGINS_PATH)
 
         # Register AYON event for workfiles loading.
         register_event_callback("workio.open_file", check_inventory_versions)
@@ -177,7 +171,6 @@ def add_nuke_callbacks():
 
     # set apply all workfile settings on script load and save
     nuke.addOnScriptLoad(WorkfileSettings().set_context_settings)
-
 
     if nuke_settings["dirmap"]["enabled"]:
         log.info("Added Nuke's dir-mapping callback ...")
