+++ client/ayon_core/pipeline/__init__.py
@@ -97,6 +97,15 @@ from .context_tools import (
     get_current_folder_path,
     get_current_task_name
 )
+
+from .workfile import (
+    discover_template_placeholder_plugins,
+    register_template_placeholder_plugin,
+    deregister_template_placeholder_plugin,
+    register_template_placeholder_plugin_path,
+    deregister_template_placeholder_plugin_path,
+)
+
 install = install_host
 uninstall = uninstall_host
 
@@ -197,6 +206,13 @@ __all__ = (
     "get_current_project_name",
     "get_current_folder_path",
     "get_current_task_name",
+
+    # Workfile templates
+    "discover_template_placeholder_plugins",
+    "register_template_placeholder_plugin",
+    "deregister_template_placeholder_plugin",
+    "register_template_placeholder_plugin_path",
+    "deregister_template_placeholder_plugin_path",
 
     # Backwards compatible function names
     "install",
