+++ client/ayon_core/pipeline/workfile/__init__.py
@@ -21,6 +21,15 @@ from .utils import (
 from .build_workfile import BuildWorkfile
 
 
+from .workfile_template_builder import (
+    discover_template_placeholder_plugins,
+    register_template_placeholder_plugin,
+    deregister_template_placeholder_plugin,
+    register_template_placeholder_plugin_path,
+    deregister_template_placeholder_plugin_path,
+)
+
+
 __all__ = (
     "get_workfile_template_key_from_context",
     "get_workfile_template_key",
@@ -39,4 +48,10 @@ __all__ = (
     "should_open_workfiles_tool_on_launch",
 
     "BuildWorkfile",
+
+    "discover_template_placeholder_plugins",
+    "register_template_placeholder_plugin",
+    "deregister_template_placeholder_plugin",
+    "register_template_placeholder_plugin_path",
+    "deregister_template_placeholder_plugin_path",
 )
