+++ client/ayon_core/pipeline/workfile/workfile_template_builder.py
@@ -43,6 +43,13 @@ from ayon_core.pipeline.load import (
     get_representation_contexts,
     load_with_repre_context,
 )
+from ayon_core.pipeline.plugin_discover import (
+    discover,
+    register_plugin,
+    register_plugin_path,
+    deregister_plugin,
+    deregister_plugin_path
+)
 
 from ayon_core.pipeline.create import (
     discover_legacy_creator_plugins,
@@ -211,10 +218,14 @@ class AbstractTemplateBuilder(object):
         Returns:
             List[PlaceholderPlugin]: Plugin classes available for host.
         """
+        plugins = []
 
+        # Backwards compatibility
         if hasattr(self._host, "get_workfile_build_placeholder_plugins"):
             return self._host.get_workfile_build_placeholder_plugins()
-        return []
+
+        plugins.extend(discover(PlaceholderPlugin))
+        return plugins
 
     @property
     def host(self):
@@ -1918,3 +1929,23 @@ class CreatePlaceholderItem(PlaceholderItem):
 
     def create_failed(self, creator_data):
         self._failed_created_publish_instances.append(creator_data)
+
+
+def discover_template_placeholder_plugins(*args, **kwargs):
+    return discover(PlaceholderPlugin, *args, **kwargs)
+
+
+def register_template_placeholder_plugin(plugin: PlaceholderPlugin):
+    register_plugin(PlaceholderPlugin, plugin)
+
+
+def deregister_template_placeholder_plugin(plugin: PlaceholderPlugin):
+    deregister_plugin(PlaceholderPlugin, plugin)
+
+
+def register_template_placeholder_plugin_path(path: str):
+    register_plugin_path(PlaceholderPlugin, path)
+
+
+def deregister_template_placeholder_plugin_path(path: str):
+    deregister_plugin_path(PlaceholderPlugin, path)
